Add an example of a Quarto presentation
I have put it in there.  It seems to work, but I just don't know if GitHub Actions will have the resources to make it work.  I hope that the quarto package will trigger download of the necessary tools, but I don't know if that will be the case.

Gonna commit and push and see!